pyproject.toml
[project]
name = "instaryt-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cloudinary>=1.44.1",
    "fastapi>=0.135.3",
    "ngrok>=1.7.0",
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "requests>=2.33.1",
    "uvicorn>=0.42.0",
]
